updated to godot 4.2.1, fixed enum size issue

+++ src/gdextension_api_ClockDirection.v
@@ -1,6 +1,6 @@
 module vgdextension
 
-pub enum ClockDirection {
+pub enum ClockDirection as i64 {
     clockwise = 0
     counterclockwise = 1
 }
