+++ src/gdextension_api_Corner.v
@@ -1,6 +1,6 @@
 module vgdextension
 
-pub enum Corner {
+pub enum Corner as i64 {
     corner_top_left = 0
     corner_top_right = 1
     corner_bottom_right = 2
