+++ src/gdextension_api_Glyph.v
@@ -2,8 +2,8 @@ module vgdextension
 
 pub struct Glyph {
     pub mut:
-    start i32 = -1
-    end i32 = -1
+    start i64 = -1
+    end i64 = -1
     count u8
     repeat u8 = 1
     flags u16
@@ -11,7 +11,7 @@ pub struct Glyph {
     y_off f64
     advance f64
     font_rid RID
-    font_size i32
+    font_size i64
     index i32
 }
