+++ src/gdextension_api_HorizontalAlignment.v
@@ -1,6 +1,6 @@
 module vgdextension
 
-pub enum HorizontalAlignment {
+pub enum HorizontalAlignment as i64 {
     horizontal_alignment_left = 0
     horizontal_alignment_center = 1
     horizontal_alignment_right = 2
