+++ src/gdextension_api_ImageFormatLoader.v
@@ -1,6 +1,6 @@
 module vgdextension
 
-pub enum ImageFormatLoaderLoaderFlags {
+pub enum ImageFormatLoaderLoaderFlags as i64 {
     flag_none = 0
     flag_force_linear = 1
     flag_convert_colors = 2
