+++ src/gdextension_api_Orientation.v
@@ -1,6 +1,6 @@
 module vgdextension
 
-pub enum Orientation {
+pub enum Orientation as i64 {
     vertical = 1
     horizontal = 0
 }
