+++ src/gdextension_api_PhysicsServer2DExtensionMotionResult.v
@@ -10,9 +10,9 @@ pub struct PhysicsServer2DExtensionMotionResult {
     collision_depth f64
     collision_safe_fraction f64
     collision_unsafe_fraction f64
-    collision_local_shape i32
+    collision_local_shape i64
     collider_id ObjectID
     collider RID
-    collider_shape i32
+    collider_shape i64
 }
