+++ src/gdextension_api_ResourceImporter.v
@@ -1,6 +1,6 @@
 module vgdextension
 
-pub enum ResourceImporterImportOrder {
+pub enum ResourceImporterImportOrder as i64 {
     import_order_default = 0
     import_order_scene = 100
 }
