+++ src/gdextension_api_Side.v
@@ -1,6 +1,6 @@
 module vgdextension
 
-pub enum Side {
+pub enum Side as i64 {
     side_left = 0
     side_top = 1
     side_right = 2
