+++ src/gdextension_api_VerticalAlignment.v
@@ -1,6 +1,6 @@
 module vgdextension
 
-pub enum VerticalAlignment {
+pub enum VerticalAlignment as i64 {
     vertical_alignment_top = 0
     vertical_alignment_center = 1
     vertical_alignment_bottom = 2
